fix: turned volume getters from double to int

+++ src/main/java/it/unibo/crabinv/Model/audio/SoundService.java
@@ -7,12 +7,12 @@ public interface SoundService {
     void playBGM(String musicName);
     void stopBGM();
     void setBGMVolume(int volume);
-    double getBGMVolume();
+    int getBGMVolume();
     void toggleMuteBGM();
     boolean isBGMMuted();
     void playSfx(String effectName);
     void setSFXVolume(int volume);
-    double getSFXVolume();
+    int getSFXVolume();
     void toggleMuteSFX();
     boolean isSFXMuted();
 }
